Use Helpers::random<std::size_t>() instead of Helpers::random<int>() in getRandomItemIndexFromContainer() function

+++ Osiris/InventoryChanger/ItemGenerator.cpp
@@ -34,7 +34,7 @@ using StaticData::TournamentMap;
 [[nodiscard]] StaticData::ItemIndex getRandomItemIndexFromContainer(const StaticData::Case& container) noexcept
 {
     assert(container.hasLoot());
-    return StaticData::caseLoot()[Helpers::random(static_cast<int>(container.lootBeginIdx), static_cast<int>(container.lootEndIdx - 1))];
+    return StaticData::caseLoot()[Helpers::random(container.lootBeginIdx, container.lootEndIdx - 1)];
 }
 
 std::pair<StaticData::ItemIndex, std::size_t> ItemGenerator::generateItemFromContainer(const InventoryItem& caseItem) noexcept
